refactor(code_gen): extract Value enum into value.rs
